tidy up

+++ src/Applications/use_case/threads/DeleteCommentsUseCase.js
@@ -4,6 +4,8 @@ class DeleteCommentsUseCase {
   }
 
   async execute (owner, threadId, commentId) {
+    await this._threadRepository.verifyThreadAvailability(threadId)
+    await this._threadRepository.verifyCommentOwner(commentId, owner)
     return await this._threadRepository.deleteThreadComments(owner, threadId, commentId)
   }
 }
